refactor: extract game store
