+++ src/components/GameController.tsx
@@ -1,1014 +1,53 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { Tile, PlayerState, LogEntry, MeldType } from '../types/mahjong';
-import { generateTileWall, drawDoraIndicator } from './TileWall';
-import {
-  createInitialPlayerState,
-  drawTiles,
-  discardTile,
-  claimMeld,
-  declareRiichi,
-  clearIppatsu,
-  isTenpaiAfterDiscard,
-  canDiscardTile,
-  canCallMeld,
-  removeDiscardTile,
-} from './Player';
-import {
-  selectMeldTiles,
-  getValidCallOptions,
-  getSelfKanOptions,
-  getChiOptions,
-} from '../utils/meld';
-import { filterChiOptions } from '../utils/table';
-import { isWinningHand, detectYaku } from '../score/yaku';
-import { calculateScore } from '../score/score';
+import React from 'react';
 import { UIBoard } from './UIBoard';
-import { calcShanten } from '../utils/shanten';
-import { incrementDiscardCount, findRonWinner } from './DiscardUtil';
-import { chooseAICallOption } from '../utils/ai';
-import { payoutTsumo, payoutRon, payoutNoten } from '../utils/payout';
-import { RoundResultModal, RoundResult } from './RoundResultModal';
+import { RoundResultModal } from './RoundResultModal';
 import { FinalResultModal } from './FinalResultModal';
-import { WinResultModal, WinResult } from './WinResultModal';
-import { exportLqRecord } from '../utils/paifuExport';
-import type { RecordHead } from '../types/jantama';
-
-const DEAD_WALL_SIZE = 14;
-
-export type GameLength = 'east1' | 'tonpu' | 'tonnan';
-export const maxKyokuForLength = (len: GameLength): number =>
-  len === 'tonnan' ? 8 : len === 'tonpu' ? 4 : 1;
-
-type GamePhase = 'init' | 'playing' | 'end';
-
-interface BoardData {
-  players: PlayerState[];
-  wall: Tile[];
-  deadWall: Tile[];
-  dora: Tile[];
-  turn: number;
-  kyoku: number;
-  riichiPool: number;
-  honba: number;
-}
-
-const boardPresets: Record<string, BoardData> = (() => {
-  // eslint-disable-next-line no-unused-vars
-  type TileFactory = (suit: Tile['suit'], rank: number) => Tile;
-  // eslint-disable-next-line no-unused-vars
-  const make = (build: (t: TileFactory) => BoardData): BoardData => {
-    let id = 1;
-    const t = (suit: Tile['suit'], rank: number): Tile => ({ suit, rank, id: `p${id++}` });
-    return build(t);
-  };
-
-  const basic = make(t => {
-    const p0 = createInitialPlayerState('自分', false, 0);
-    const chi = [t('man', 1), t('man', 2), t('man', 3)];
-    p0.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('sou', 1), t('sou', 2), t('sou', 3), t('wind', 1), t('wind', 2), t('dragon', 1), t('dragon', 2), t('man', 7), t('man', 8), t('man', 9)];
-    p0.melds = [{ type: 'chi', tiles: chi, fromPlayer: 1, calledTileId: chi[0].id }];
-
-    const p1 = createInitialPlayerState('下家', true, 1);
-    const pon = [t('pin', 7), t('pin', 7), t('pin', 7)];
-    p1.hand = [t('man', 1), t('man', 2), t('man', 3), t('sou', 4), t('sou', 5), t('sou', 6), t('wind', 3), t('wind', 4), t('dragon', 3), t('man', 5), t('pin', 9), t('pin', 9)];
-    p1.melds = [{ type: 'pon', tiles: pon, fromPlayer: 2, calledTileId: pon[1].id }];
-
-    const p2 = createInitialPlayerState('対面', true, 2);
-    const kan = [t('sou', 1), t('sou', 1), t('sou', 1), t('sou', 1)];
-    p2.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('man', 4), t('man', 5), t('man', 6), t('wind', 1), t('wind', 2), t('dragon', 1)];
-    p2.melds = [{ type: 'kan', tiles: kan, fromPlayer: 3, calledTileId: kan[0].id }];
-
-    const p3 = createInitialPlayerState('上家', true, 3);
-    p3.hand = [t('man', 2), t('pin', 2), t('sou', 2), t('wind', 1), t('wind', 2), t('wind', 3), t('dragon', 1), t('dragon', 2), t('dragon', 3), t('man', 9), t('pin', 9), t('sou', 9), t('man', 1)];
-
-    return { players: [p0, p1, p2, p3], wall: [], deadWall: [], dora: [], turn: 0, kyoku: 1, riichiPool: 0, honba: 0 };
-  });
-
-  const multiCalls = make(t => {
-    const p0 = createInitialPlayerState('自分', false, 0);
-    p0.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('sou', 1), t('sou', 2), t('sou', 3), t('wind', 1), t('wind', 2), t('dragon', 1), t('dragon', 2), t('man', 7), t('man', 8), t('man', 9)];
-
-    const p1 = createInitialPlayerState('下家', true, 1);
-    const pon = [t('pin', 7), t('pin', 7), t('pin', 7)];
-    const chi = [t('man', 1), t('man', 2), t('man', 3)];
-    p1.hand = [t('sou', 4), t('sou', 5), t('sou', 6), t('wind', 3), t('wind', 4), t('dragon', 3), t('man', 5), t('pin', 9), t('pin', 9)];
-    p1.melds = [
-      { type: 'pon', tiles: pon, fromPlayer: 2, calledTileId: pon[1].id },
-      { type: 'chi', tiles: chi, fromPlayer: 0, calledTileId: chi[0].id },
-    ];
-
-    const p2 = createInitialPlayerState('対面', true, 2);
-    const kan = [t('sou', 1), t('sou', 1), t('sou', 1), t('sou', 1)];
-    p2.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('man', 4), t('man', 5), t('man', 6), t('wind', 1), t('wind', 2), t('dragon', 1)];
-    p2.melds = [{ type: 'kan', tiles: kan, fromPlayer: 3, calledTileId: kan[0].id }];
-
-    const p3 = createInitialPlayerState('上家', true, 3);
-    p3.hand = [t('man', 2), t('pin', 2), t('sou', 2), t('wind', 1), t('wind', 2), t('wind', 3), t('dragon', 1), t('dragon', 2), t('dragon', 3), t('man', 9), t('pin', 9), t('sou', 9), t('man', 1)];
-
-    return { players: [p0, p1, p2, p3], wall: [], deadWall: [], dora: [], turn: 0, kyoku: 1, riichiPool: 0, honba: 0 };
-  });
-
-  const kanVariants = make(t => {
-    const p0 = createInitialPlayerState('自分', false, 0);
-    const ankan = [t('man', 5), t('man', 5), t('man', 5), t('man', 5)];
-    p0.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('sou', 1), t('sou', 2), t('sou', 3), t('wind', 1), t('wind', 2), t('dragon', 1), t('dragon', 2), t('man', 7), t('man', 8), t('man', 9)];
-    p0.melds = [{ type: 'kan', tiles: ankan, fromPlayer: 0, calledTileId: ankan[0].id }];
-
-    const p1 = createInitialPlayerState('下家', true, 1);
-    const minkan = [t('pin', 7), t('pin', 7), t('pin', 7), t('pin', 7)];
-    p1.hand = [t('man', 1), t('man', 2), t('man', 3), t('sou', 4), t('sou', 5), t('sou', 6), t('wind', 3), t('wind', 4), t('dragon', 3)];
-    p1.melds = [{ type: 'kan', tiles: minkan, fromPlayer: 2, calledTileId: minkan[1].id }];
-
-    const p2 = createInitialPlayerState('対面', true, 2);
-    const kakan = [t('sou', 3), t('sou', 3), t('sou', 3), t('sou', 3)];
-    p2.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('man', 4), t('man', 5), t('man', 6), t('wind', 1), t('wind', 2), t('dragon', 1)];
-    p2.melds = [{ type: 'kan', tiles: kakan, fromPlayer: 1, calledTileId: kakan[0].id }];
-
-    const p3 = createInitialPlayerState('上家', true, 3);
-    p3.hand = [t('man', 2), t('pin', 2), t('sou', 2), t('wind', 1), t('wind', 2), t('wind', 3), t('dragon', 1), t('dragon', 2), t('dragon', 3), t('man', 9), t('pin', 9), t('sou', 9), t('man', 1)];
-
-    return { players: [p0, p1, p2, p3], wall: [], deadWall: [], dora: [], turn: 0, kyoku: 1, riichiPool: 0, honba: 0 };
-  });
-
-  const longRiver = make(t => {
-    const p0 = createInitialPlayerState('自分', false, 0);
-    const chi = [t('man', 1), t('man', 2), t('man', 3)];
-    p0.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('sou', 1), t('sou', 2), t('sou', 3), t('wind', 1), t('wind', 2), t('dragon', 1), t('dragon', 2), t('man', 7), t('man', 8), t('man', 9)];
-    p0.melds = [{ type: 'chi', tiles: chi, fromPlayer: 1, calledTileId: chi[0].id }];
-
-    const p1 = createInitialPlayerState('下家', true, 1);
-    const pon = [t('pin', 7), t('pin', 7), t('pin', 7)];
-    p1.hand = [t('man', 1), t('man', 2), t('man', 3), t('sou', 4), t('sou', 5), t('sou', 6), t('wind', 3), t('wind', 4), t('dragon', 3), t('man', 5), t('pin', 9), t('pin', 9)];
-    p1.melds = [{ type: 'pon', tiles: pon, fromPlayer: 2, calledTileId: pon[1].id }];
-
-    const p2 = createInitialPlayerState('対面', true, 2);
-    const kan = [t('sou', 1), t('sou', 1), t('sou', 1), t('sou', 1)];
-    p2.hand = [t('pin', 1), t('pin', 2), t('pin', 3), t('man', 4), t('man', 5), t('man', 6), t('wind', 1), t('wind', 2), t('dragon', 1)];
-    p2.melds = [{ type: 'kan', tiles: kan, fromPlayer: 3, calledTileId: kan[0].id }];
-
-    const p3 = createInitialPlayerState('上家', true, 3);
-    p3.hand = [t('man', 2), t('pin', 2), t('sou', 2), t('wind', 1), t('wind', 2), t('wind', 3), t('dragon', 1), t('dragon', 2), t('dragon', 3), t('man', 9), t('pin', 9), t('sou', 9), t('man', 1)];
-    p3.discard = Array.from({ length: 20 }).map((_, i) => t('man', (i % 9) + 1));
-
-    return { players: [p0, p1, p2, p3], wall: [], deadWall: [], dora: [], turn: 0, kyoku: 1, riichiPool: 0, honba: 0 };
-  });
-
-  return { basic, multiCalls, kanVariants, longRiver };
-})();
+import { WinResultModal } from './WinResultModal';
+import {
+  useGame,
+  maxKyokuForLength,
+  GameLength,
+} from '../game/store';
 
 interface Props {
   gameLength: GameLength;
 }
 
 export const GameController: React.FC<Props> = ({ gameLength }) => {
-  // ゲーム状態
-  const [wall, setWall] = useState<Tile[]>([]);
-  const [players, setPlayers] = useState<PlayerState[]>([]);
-  const [dora, setDora] = useState<Tile[]>([]);
-  const [deadWall, setDeadWall] = useState<Tile[]>([]);
-  const [playerIsAI, setPlayerIsAI] = useState(false);
-  const [turn, setTurn] = useState(0); // 0:自分, 1-3:AI
-  const [phase, setPhase] = useState<GamePhase>('init');
-  const [message, setMessage] = useState<string>('');
-  const [kyoku, setKyoku] = useState<number>(1); // 東1局など
-  const [shanten, setShanten] = useState<{ standard: number; chiitoi: number; kokushi: number }>({ standard: 8, chiitoi: 8, kokushi: 13 });
-  const [discardCounts, setDiscardCounts] = useState<Record<string, number>>({});
-  const [lastDiscard, setLastDiscard] = useState<{ tile: Tile; player: number; isShonpai: boolean } | null>(null);
-  const [callOptions, setCallOptions] = useState<(MeldType | 'pass')[] | null>(null);
-  const [roundResult, setRoundResult] = useState<RoundResult | null>(null);
-  const [winResult, setWinResult] = useState<WinResult | null>(null);
-  const [selfKanOptions, setSelfKanOptions] = useState<Tile[][] | null>(null);
-  const [chiTileOptions, setChiTileOptions] = useState<Tile[][] | null>(null);
-  const [riichiPool, setRiichiPool] = useState(0);
-  const [honba, setHonba] = useState(0);
-  const [pendingRiichi, setPendingRiichi] = useState<number | null>(null);
-  const [tsumoOption, setTsumoOption] = useState(false);
-  const [ronCandidate, setRonCandidate] = useState<{ tile: Tile; from: number } | null>(null);
-  const [log, setLog] = useState<LogEntry[]>([]);
-  const [boardInput, setBoardInput] = useState('');
-
-  const turnRef = useRef(turn);
-  const playersRef = useRef<PlayerState[]>(players);
-  const wallRef = useRef<Tile[]>(wall);
-  const deadWallRef = useRef<Tile[]>(deadWall);
-  const kyokuRef = useRef(kyoku);
-  const riichiPoolRef = useRef(riichiPool);
-  const honbaRef = useRef(honba);
-  const logRef = useRef<LogEntry[]>(log);
-  const kanDrawRef = useRef<number | null>(null);
-  const drawInfoRef = useRef<Record<number, { rinshan: boolean; last: boolean }>>({});
-  const recordHeadRef = useRef<RecordHead>({
-    startTime: 0,
-    endTime: 0,
-    rule: { gameLength },
-    players: [],
-  });
-  const [preset, setPreset] = useState<keyof typeof boardPresets>('basic');
-
-  useEffect(() => {
-    setBoardInput(JSON.stringify(boardPresets[preset], null, 2));
-  }, [preset]);
-
-  const togglePlayerAI = () => {
-    setPlayerIsAI(prev => {
-      const next = !prev;
-      setPlayers(ps =>
-        ps.map((pl, idx) => (idx === 0 ? { ...pl, isAI: next } : pl)),
-      );
-      playersRef.current = playersRef.current.map((pl, idx) =>
-        idx === 0 ? { ...pl, isAI: next } : pl,
-      );
-      if (next && turnRef.current === 0) {
-        setTimeout(() => {
-          const tile = playersRef.current[0].hand[0];
-          handleDiscard(tile.id);
-        }, 500);
-      }
-      return next;
-    });
-  };
-
-  useEffect(() => {
-    turnRef.current = turn;
-  }, [turn]);
-
-  useEffect(() => {
-    wallRef.current = wall;
-  }, [wall]);
-
-  useEffect(() => {
-    deadWallRef.current = deadWall;
-  }, [deadWall]);
-
-  useEffect(() => {
-    kyokuRef.current = kyoku;
-  }, [kyoku]);
-
-  useEffect(() => {
-    riichiPoolRef.current = riichiPool;
-  }, [riichiPool]);
-
-  useEffect(() => {
-    honbaRef.current = honba;
-  }, [honba]);
-
-  useEffect(() => {
-    logRef.current = log;
-  }, [log]);
-
-  useEffect(() => {
-    playersRef.current = players;
-    if (players.length > 0) {
-      setShanten(calcShanten(players[0].hand, players[0].melds.length));
-    }
-  }, [players]);
-
-  // ラウンド初期化関数
-  const startRound = (resetKyoku: boolean, roundNumber: number = kyokuRef.current) => {
-    let wallStack = generateTileWall();
-    let wanpai = wallStack.slice(0, DEAD_WALL_SIZE);
-    wallStack = wallStack.slice(DEAD_WALL_SIZE);
-    const doraResult = drawDoraIndicator(wanpai, 1);
-    const doraTiles = doraResult.dora;
-    wanpai = doraResult.wall;
-    let p: PlayerState[];
-    if (resetKyoku) {
-      p = [
-        createInitialPlayerState('あなた', playerIsAI, 0),
-        createInitialPlayerState('AI下家', true, 1),
-        createInitialPlayerState('AI対面', true, 2),
-        createInitialPlayerState('AI上家', true, 3),
-      ];
-    } else {
-      p = playersRef.current.map(pl => ({
-        ...pl,
-        hand: [],
-        discard: [],
-        melds: [],
-        drawnTile: null,
-        isRiichi: false,
-        isAI: pl.seat === 0 ? playerIsAI : pl.isAI,
-      }));
-    }
-    for (let i = 0; i < 4; i++) {
-      const result = drawTiles(p[i], wallStack, 13);
-      p[i] = result.player;
-      wallStack = result.wall;
-    }
-    // 親は1枚多く持つ
-    const extra = drawTiles(p[0], wallStack, 1);
-    p[0] = extra.player;
-    wallStack = extra.wall;
-    setPlayers(p);
-    playersRef.current = p;
-    setWall(wallStack);
-    wallRef.current = wallStack;
-    setDeadWall(wanpai);
-    deadWallRef.current = wanpai;
-    setDora(doraTiles);
-    setTurn(0);
-    setDiscardCounts({});
-    setLastDiscard(null);
-    setPendingRiichi(null);
-    setTsumoOption(false);
-    setRonCandidate(null);
-    setRoundResult(null);
-    setWinResult(null);
-    if (resetKyoku) {
-      setRiichiPool(0);
-      setHonba(0);
-      honbaRef.current = 0;
-      recordHeadRef.current.startTime = Date.now();
-      recordHeadRef.current.endTime = 0;
-      recordHeadRef.current.rule = { gameLength };
-      recordHeadRef.current.players = playersRef.current.map(p => ({
-        name: p.name,
-        seat: p.seat,
-        isAI: p.isAI,
-      }));
-    }
-    setMessage(
-      `配牌が完了しました。${playerIsAI ? 'AIのターンです。' : 'あなたのターンです。'}`,
-    );
-    setLog([{ type: 'startRound', kyoku: roundNumber }]);
-    logRef.current = [{ type: 'startRound', kyoku: roundNumber }];
-    kanDrawRef.current = null;
-    drawInfoRef.current = {};
-    setPhase('playing');
-  };
-
-  // 初期化
-  useEffect(() => {
-    if (phase === 'init') {
-      setKyoku(1);
-      startRound(true, 1);
-    }
-  }, [phase]);
-
-  useEffect(() => {
-    if (phase === 'end') {
-      recordHeadRef.current.endTime = Date.now();
-    }
-  }, [phase]);
-
+  const game = useGame(gameLength);
   const maxKyoku = maxKyokuForLength(gameLength);
-  const nextKyoku = (dealerContinues = false) => {
-    if (dealerContinues) {
-      setHonba(h => h + 1);
-      honbaRef.current += 1;
-      startRound(false, kyokuRef.current);
-      return;
-    }
-    setHonba(0);
-    honbaRef.current = 0;
-    const next = kyokuRef.current + 1;
-    if (next > maxKyoku) {
-      setPhase('end');
-    } else {
-      setKyoku(next);
-      startRound(false, next);
-    }
-  };
 
-  const handleWallExhaustion = () => {
-    const tenpai = playersRef.current.map(p => {
-      const s = calcShanten(p.hand, p.melds.length);
-      return Math.min(s.standard, s.chiitoi, s.kokushi) === 0;
-    });
-    const { players: updated, changes } = payoutNoten(playersRef.current, tenpai);
-    setPlayers(updated);
-    playersRef.current = updated;
-    const results: RoundResult = {
-      results: updated.map((p, idx) => ({
-        name: p.name,
-        score: p.score,
-        change: changes[idx],
-        isTenpai: tenpai[idx],
-      })),
-    };
-    setRoundResult(results);
-    setMessage('牌山が尽きました。流局です。');
-  };
-
-  // ツモ処理
-  const drawForCurrentPlayer = () => {
-    if (wallRef.current.length === 0) {
-      handleWallExhaustion();
-      return;
-    }
-    const currentIndex = turnRef.current;
-    let p = [...playersRef.current];
-    const result = drawTiles(p[currentIndex], wallRef.current, 1);
-    p[currentIndex] = result.player;
-    setPlayers(p);
-    playersRef.current = p;
-    setWall(result.wall);
-    wallRef.current = result.wall;
-    const rinshan = kanDrawRef.current === currentIndex;
-    if (rinshan) kanDrawRef.current = null;
-    const last = result.wall.length === 0;
-    drawInfoRef.current[currentIndex] = { rinshan, last };
-    setLog(prev => [...prev, { type: 'draw', player: currentIndex, tile: result.player.drawnTile as Tile }]);
-    logRef.current = [...logRef.current, { type: 'draw', player: currentIndex, tile: result.player.drawnTile as Tile }];
-    if (last) {
-      // postpone exhaustion until after possible win/discard
-    }
-    if (!playersRef.current[currentIndex].isAI) {
-      const opts = getSelfKanOptions(playersRef.current[currentIndex]);
-      setSelfKanOptions(opts.length > 0 ? opts : null);
-    } else {
-      const opts = getSelfKanOptions(playersRef.current[currentIndex]);
-      if (opts.length > 0) {
-        performSelfKan(currentIndex, opts[0]);
-        return;
-      }
-    }
-    if (isWinningHand([...p[currentIndex].hand, ...p[currentIndex].melds.flatMap(m => m.tiles)])) {
-      if (p[currentIndex].isAI || currentIndex !== 0) {
-        performTsumo(currentIndex);
-        return;
-      } else {
-        setTsumoOption(true);
-        return;
-      }
-    }
-    if (last) {
-      handleWallExhaustion();
-      return;
-    }
-
-  };
-
-  // 捨て牌処理（自分／AI共通）
-  const handleDiscard = (tileId: string) => {
-    const idx = turnRef.current;
-    let p = [...playersRef.current];
-    const tile = p[idx].hand.find(t => t.id === tileId);
-    if (!tile) return;
-    if (pendingRiichi !== idx && !canDiscardTile(p[idx], tileId)) {
-      setMessage('リーチ後はツモ牌しか切れません');
-      return;
-    }
-    if (pendingRiichi === idx && !isTenpaiAfterDiscard(p[idx], tileId)) {
-      setMessage('その牌ではリーチできません');
-      return;
-    }
-    setSelfKanOptions(null);
-    setChiTileOptions(null);
-    if (pendingRiichi !== idx && p[idx].ippatsu) {
-      p[idx] = clearIppatsu(p[idx]);
-    }
-    const result = incrementDiscardCount(discardCounts, tile);
-    setDiscardCounts(result.record);
-    setLastDiscard({ tile, player: idx, isShonpai: result.isShonpai });
-    const isRiichi = p[idx].isRiichi;
-    p[idx] = discardTile(p[idx], tileId, isRiichi);
-    setPlayers(p);
-    playersRef.current = p;
-    setLog(prev => [...prev, { type: 'discard', player: idx, tile }]);
-    logRef.current = [...logRef.current, { type: 'discard', player: idx, tile }];
-    const winIdx = findRonWinner(p, idx, tile);
-    if (winIdx !== null) {
-      if (p[winIdx].isAI || winIdx !== 0) {
-        performRon(winIdx, idx, tile);
-        return;
-      } else {
-        setRonCandidate({ tile, from: idx });
-        return;
-      }
-    }
-    if (pendingRiichi === idx) {
-      p[idx] = { ...p[idx], score: p[idx].score - 1000, ippatsu: true };
-      setRiichiPool(prev => prev + 1);
-      setPendingRiichi(null);
-      setPlayers(p);
-      playersRef.current = p;
-    }
-    const info = drawInfoRef.current[idx];
-    drawInfoRef.current[idx] = { rinshan: false, last: false };
-    if (info?.last) {
-      handleWallExhaustion();
-      return;
-    }
-    if (idx !== 0 && !playersRef.current[0].isAI) {
-      let options = getValidCallOptions(p[0], tile);
-      options = filterChiOptions(
-        options,
-        playersRef.current[0].seat,
-        playersRef.current[idx].seat,
-      );
-      const hasAction = options.some(o => o !== 'pass');
-      if (!hasAction) {
-        setCallOptions(null);
-        setLastDiscard(null);
-        nextTurn();
-      } else {
-        setCallOptions(options);
-      }
-    } else {
-      nextTurn();
-    }
-  };
-
-
-const handleCallAction = (action: MeldType | 'pass') => {
-  if (!lastDiscard) return;
-  if (!canCallMeld(playersRef.current[0])) {
-    setCallOptions(null);
-    setLastDiscard(null);
-    return;
-  }
-  if (action === 'pass') {
-    setCallOptions(null);
-    setLastDiscard(null);
-    setSelfKanOptions(null);
-    setChiTileOptions(null);
-    nextTurn();
-    return;
-  }
-  const caller = 0;
-  const discarder = lastDiscard.player;
-  let p = [...playersRef.current];
-  p = p.map(pl => clearIppatsu(pl));
-  if (action === 'chi') {
-    const options = getChiOptions(p[caller], lastDiscard.tile);
-    if (options.length > 1 && !chiTileOptions) {
-      setChiTileOptions(options);
-      setCallOptions(['pass']);
-      return;
-    }
-  }
-  const meldTiles = selectMeldTiles(p[caller], lastDiscard.tile, action);
-  if (!meldTiles) {
-    setCallOptions(null);
-    setLastDiscard(null);
-    nextTurn();
-    return;
-    }
-  p[discarder] = removeDiscardTile(p[discarder], lastDiscard.tile.id);
-  p[caller] = claimMeld(
-    p[caller],
-    [...meldTiles, lastDiscard.tile],
-    action,
-    discarder,
-    lastDiscard.tile.id,
-  );
-  setPlayers(p);
-  playersRef.current = p;
-  setLog(prev => [
-    ...prev,
-    {
-      type: 'meld',
-      player: caller,
-      tiles: [...meldTiles, lastDiscard.tile],
-      meldType: action,
-      from: discarder,
-    },
-  ]);
-  logRef.current = [
-    ...logRef.current,
-    {
-      type: 'meld',
-      player: caller,
-      tiles: [...meldTiles, lastDiscard.tile],
-      meldType: action,
-      from: discarder,
-    },
-  ];
-
-  if (action === 'kan') {
-    kanDrawRef.current = caller;
-  }
-
-  if (action === 'kan') {
-    const doraResult = drawDoraIndicator(deadWallRef.current, 1);
-    setDora(prev => [...prev, ...doraResult.dora]);
-    setDeadWall(doraResult.wall);
-    deadWallRef.current = doraResult.wall;
-      turnRef.current = caller;
-      drawForCurrentPlayer();
-    }
-
-  setCallOptions(null);
-  setLastDiscard(null);
-  setSelfKanOptions(null);
-  setChiTileOptions(null);
-  setTurn(caller);
-};
-
-  const performSelfKan = (caller: number, tiles: Tile[]) => {
-    if (!canCallMeld(playersRef.current[caller])) return;
-    let p = playersRef.current.map(pl => clearIppatsu(pl));
-  p = [...p];
-  p[caller] = claimMeld(p[caller], tiles, 'kan', caller, tiles[0].id);
-  setPlayers(p);
-  playersRef.current = p;
-  setLog(prev => [
-    ...prev,
-    { type: 'meld', player: caller, tiles, meldType: 'kan', from: caller },
-  ]);
-  logRef.current = [
-    ...logRef.current,
-    { type: 'meld', player: caller, tiles, meldType: 'kan', from: caller },
-  ];
-
-  kanDrawRef.current = caller;
-
-  const doraResult = drawDoraIndicator(deadWallRef.current, 1);
-    setDora(prev => [...prev, ...doraResult.dora]);
-    setDeadWall(doraResult.wall);
-    deadWallRef.current = doraResult.wall;
-    turnRef.current = caller;
-    drawForCurrentPlayer();
-  };
-
-  const performAICall = (caller: number, action: MeldType) => {
-    if (!lastDiscard) return;
-    if (!canCallMeld(playersRef.current[caller])) return;
-    const discarder = lastDiscard.player;
-    let p = playersRef.current.map(pl => clearIppatsu(pl));
-    p = [...p];
-    const meldTiles = selectMeldTiles(p[caller], lastDiscard.tile, action);
-    if (!meldTiles) return;
-    p[discarder] = removeDiscardTile(p[discarder], lastDiscard.tile.id);
-    p[caller] = claimMeld(
-      p[caller],
-      [...meldTiles, lastDiscard.tile],
-      action,
-      discarder,
-      lastDiscard.tile.id,
-    );
-    setPlayers(p);
-    playersRef.current = p;
-    setLog(prev => [
-      ...prev,
-      {
-        type: 'meld',
-        player: caller,
-        tiles: [...meldTiles, lastDiscard.tile],
-        meldType: action,
-        from: discarder,
-      },
-    ]);
-    logRef.current = [
-      ...logRef.current,
-      {
-        type: 'meld',
-        player: caller,
-        tiles: [...meldTiles, lastDiscard.tile],
-        meldType: action,
-        from: discarder,
-      },
-    ];
-    setMessage(`${p[caller].name} が ${action}しました。`);
-
-    if (action === 'kan') {
-      const doraResult = drawDoraIndicator(deadWallRef.current, 1);
-      setDora(prev => [...prev, ...doraResult.dora]);
-      setDeadWall(doraResult.wall);
-      deadWallRef.current = doraResult.wall;
-      turnRef.current = caller;
-      drawForCurrentPlayer();
-    }
-
-    setLastDiscard(null);
-    setTurn(caller);
-    setTimeout(() => {
-      const tile = playersRef.current[caller].hand[0];
-      handleDiscard(tile.id);
-    }, 500);
-  };
-
-  const performTsumo = (idx: number) => {
-    const p = [...playersRef.current];
-    let ura: Tile[] = [];
-    if (p[idx].isRiichi) {
-      const uraRes = drawDoraIndicator(deadWallRef.current, dora.length);
-      ura = uraRes.dora;
-      setDeadWall(uraRes.wall);
-      deadWallRef.current = uraRes.wall;
-    }
-    const fullHand = [...p[idx].hand, ...p[idx].melds.flatMap(m => m.tiles)];
-    const info = drawInfoRef.current[idx];
-    drawInfoRef.current[idx] = { rinshan: false, last: false };
-    const seatWind = p[idx].seat + 1;
-    const roundWind = kyokuRef.current <= 4 ? 1 : 2;
-    const yaku = detectYaku(fullHand, p[idx].melds, {
-      isTsumo: true,
-      isRiichi: p[idx].isRiichi,
-      doubleRiichi: p[idx].doubleRiichi,
-      ippatsu: p[idx].ippatsu,
-      rinshan: info?.rinshan,
-      haitei: info?.last,
-      seatWind,
-      roundWind,
-      uraDoraIndicators: ura,
-    });
-    const { han, fu, points } = calculateScore(
-      p[idx].hand,
-      p[idx].melds,
-      yaku,
-      dora,
-      { seatWind, roundWind, winType: 'tsumo' },
-    );
-    let newPlayers = payoutTsumo(p, idx, points, honbaRef.current).map((pl, i) =>
-      i === idx ? { ...pl, ippatsu: false } : pl,
-    );
-    if (riichiPoolRef.current > 0) {
-      newPlayers = newPlayers.map((pl, i) =>
-        i === idx ? { ...pl, score: pl.score + riichiPoolRef.current * 1000 } : pl,
-      );
-      setRiichiPool(0);
-      riichiPoolRef.current = 0;
-    }
-    setPlayers(newPlayers);
-    playersRef.current = newPlayers;
-    setLog(prev => [...prev, { type: 'tsumo', player: idx, tile: p[idx].drawnTile as Tile }]);
-    logRef.current = [...logRef.current, { type: 'tsumo', player: idx, tile: p[idx].drawnTile as Tile }];
-    setMessage(`${p[idx].name} の和了！`);
-    setTsumoOption(false);
-    setWinResult({
-      players: newPlayers,
-      winner: idx,
-      winType: 'tsumo',
-      yaku: yaku.map(y => y.name),
-      han,
-      fu,
-      points,
-      uraDora: ura,
-    });
-  };
-
-  const performRon = (winner: number, from: number, tile: Tile) => {
-    const p = [...playersRef.current];
-    let ura: Tile[] = [];
-    if (p[winner].isRiichi) {
-      const uraRes = drawDoraIndicator(deadWallRef.current, dora.length);
-      ura = uraRes.dora;
-      setDeadWall(uraRes.wall);
-      deadWallRef.current = uraRes.wall;
-    }
-    const fullHand = [...p[winner].hand, ...p[winner].melds.flatMap(m => m.tiles), tile];
-    const fromInfo = drawInfoRef.current[from];
-    drawInfoRef.current[from] = { rinshan: false, last: false };
-    const prev = logRef.current[logRef.current.length - 1];
-    const chankan = prev && prev.type === 'meld' && prev.meldType === 'kan' && prev.player === from && prev.tiles.some(t => t.id === tile.id);
-    const seatWind = p[winner].seat + 1;
-    const roundWind = kyokuRef.current <= 4 ? 1 : 2;
-    const yaku = detectYaku(fullHand, p[winner].melds, {
-      isTsumo: false,
-      isRiichi: p[winner].isRiichi,
-      doubleRiichi: p[winner].doubleRiichi,
-      ippatsu: p[winner].ippatsu,
-      chankan,
-      houtei: fromInfo?.last,
-      seatWind,
-      roundWind,
-      uraDoraIndicators: ura,
-    });
-    const { han, fu, points } = calculateScore(
-      [...p[winner].hand, tile],
-      p[winner].melds,
-      yaku,
-      dora,
-      { seatWind, roundWind, winType: 'ron' },
-    );
-    let updated = payoutRon(p, winner, from, points, honbaRef.current).map((pl, i) =>
-      i === winner ? { ...pl, ippatsu: false } : pl,
-    );
-    if (riichiPoolRef.current > 0) {
-      updated = updated.map((pl, i) =>
-        i === winner ? { ...pl, score: pl.score + riichiPoolRef.current * 1000 } : pl,
-      );
-      setRiichiPool(0);
-      riichiPoolRef.current = 0;
-    }
-    setPlayers(updated);
-    playersRef.current = updated;
-    setLog(prev => [...prev, { type: 'ron', player: winner, tile, from }]);
-    logRef.current = [...logRef.current, { type: 'ron', player: winner, tile, from }];
-    setMessage(`${p[winner].name} のロン！`);
-    setWinResult({
-      players: updated,
-      winner,
-      winType: 'ron',
-      yaku: yaku.map(y => y.name),
-      han,
-      fu,
-      points,
-      uraDora: ura,
-    });
-  };
-
-  const handleRiichi = () => {
-    let p = [...playersRef.current];
-    const isDouble = playersRef.current.every(pl => pl.discard.length === 0);
-    p[0] = declareRiichi(p[0], isDouble);
-    setPlayers(p);
-    playersRef.current = p;
-    setPendingRiichi(0);
-    setMessage('リーチする牌を選んでください');
-    setLog(prev => [...prev, { type: 'riichi', player: 0, tile: p[0].drawnTile as Tile }]);
-    logRef.current = [...logRef.current, { type: 'riichi', player: 0, tile: p[0].drawnTile as Tile }];
-  };
-
-  const handleSelfKan = (tiles: Tile[]) => {
-    if (!canCallMeld(playersRef.current[0])) return;
-    setSelfKanOptions(null);
-    performSelfKan(0, tiles);
-  };
-
-  const handleChiSelect = (tiles: Tile[]) => {
-    if (!lastDiscard) return;
-    if (!canCallMeld(playersRef.current[0])) return;
-    const caller = 0;
-    const discarder = lastDiscard.player;
-    let p = [...playersRef.current];
-  p[discarder] = removeDiscardTile(p[discarder], lastDiscard.tile.id);
-  p[caller] = claimMeld(
-    p[caller],
-    [...tiles, lastDiscard.tile],
-    'chi',
-    discarder,
-    lastDiscard.tile.id,
-  );
-  setPlayers(p);
-  playersRef.current = p;
-  setLog(prev => [
-    ...prev,
-    { type: 'meld', player: caller, tiles: [...tiles, lastDiscard.tile], meldType: 'chi', from: discarder },
-  ]);
-  logRef.current = [
-    ...logRef.current,
-    { type: 'meld', player: caller, tiles: [...tiles, lastDiscard.tile], meldType: 'chi', from: discarder },
-  ];
-    setCallOptions(null);
-    setLastDiscard(null);
-    setChiTileOptions(null);
-    setSelfKanOptions(null);
-    setTurn(caller);
-  };
-
-  const handleTsumo = () => {
-    performTsumo(0);
-  };
-
-  const handleTsumoPass = () => {
-    setTsumoOption(false);
-  };
-
-  const handleRon = () => {
-    if (!ronCandidate) return;
-    const { tile, from } = ronCandidate;
-    setRonCandidate(null);
-    performRon(0, from, tile);
-  };
-
-  const handleRonPass = () => {
-    if (!ronCandidate) return;
-    const { tile, from } = ronCandidate;
-    setRonCandidate(null);
-    if (turnRef.current !== from) return; // safety
-    if (!playersRef.current[0].isAI) {
-      let options = getValidCallOptions(playersRef.current[0], tile);
-      options = filterChiOptions(
-        options,
-        playersRef.current[0].seat,
-        playersRef.current[from].seat,
-      );
-      const hasAction = options.some(o => o !== 'pass');
-      if (!hasAction) {
-        setCallOptions(null);
-        setLastDiscard(null);
-        nextTurn();
-      } else {
-        setCallOptions(options);
-      }
-    } else {
-      nextTurn();
-    }
-  };
-
-  // ターン進行
-  const nextTurn = () => {
-    setTsumoOption(false);
-    setRonCandidate(null);
-    let next = (turnRef.current + 1) % 4;
-    setTurn(next);
-    setTimeout(() => {
-      if (playersRef.current[next].isAI) {
-        // Check if the AI wants to call on the previous discard
-        if (lastDiscard && lastDiscard.player !== next) {
-          const action = chooseAICallOption(
-            playersRef.current[next],
-            lastDiscard.tile,
-          );
-          if (action !== 'pass') {
-            performAICall(next, action);
-            return;
-          }
-          setLastDiscard(null);
-        }
-        drawForCurrentPlayer();
-        if (wallRef.current.length === 0) return;
-        // AIの打牌ロジック（現時点はランダム）
-        setTimeout(() => {
-          const tile = playersRef.current[next].hand[0];
-          handleDiscard(tile.id);
-        }, 500);
-      } else {
-        drawForCurrentPlayer();
-      }
-    }, 500);
-  };
-
-  // リセット
-  const handleRestart = () => {
-    // Start a completely new game with fresh scores
-    setKyoku(1);
-    startRound(true, 1);
-  };
-
-  const handleDownloadLog = () => {
-    const record = exportLqRecord(logRef.current, recordHeadRef.current);
-    const data = JSON.stringify(record, null, 2);
-    const blob = new Blob([data], { type: 'application/json' });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = 'log.lq.GameDetailRecords';
-    a.click();
-    URL.revokeObjectURL(url);
-  };
-
-  const handleLoadBoard = () => {
-    try {
-      const data: BoardData = JSON.parse(boardInput);
-      setPlayers(data.players);
-      playersRef.current = data.players;
-      setWall(data.wall);
-      wallRef.current = data.wall;
-      setDeadWall(data.deadWall);
-      deadWallRef.current = data.deadWall;
-      setDora(data.dora);
-      setTurn(data.turn);
-      turnRef.current = data.turn;
-      setKyoku(data.kyoku);
-      kyokuRef.current = data.kyoku;
-      setRiichiPool(data.riichiPool);
-      riichiPoolRef.current = data.riichiPool;
-      setHonba(data.honba);
-      honbaRef.current = data.honba;
-      setCallOptions(null);
-      setLastDiscard(null);
-      setSelfKanOptions(null);
-      setChiTileOptions(null);
-      setPendingRiichi(null);
-      setTsumoOption(false);
-      setRonCandidate(null);
-      setRoundResult(null);
-      setWinResult(null);
-      setLog([]);
-      logRef.current = [];
-      setPhase('playing');
-      setMessage('盤面を読み込みました');
-    } catch (e) {
-      setMessage('盤面の読み込みに失敗しました');
-    }
-  };
-
-  // UI
   return (
     <div className="p-2 flex flex-col gap-4">
       <UIBoard
-        players={players}
-        dora={dora}
-        kyoku={kyoku}
-        wallCount={wall.length}
-        kyotaku={riichiPool}
-        honba={honba}
-        onDiscard={handleDiscard}
-        isMyTurn={turn === 0 && !players[0]?.isAI}
-        shanten={shanten}
-        lastDiscard={lastDiscard}
-        callOptions={!players[0]?.isAI ? callOptions ?? undefined : undefined}
-        onCallAction={!players[0]?.isAI ? handleCallAction : undefined}
-        onRiichi={!players[0]?.isAI ? handleRiichi : undefined}
-        selfKanOptions={!players[0]?.isAI ? selfKanOptions ?? undefined : undefined}
-        onSelfKan={!players[0]?.isAI ? handleSelfKan : undefined}
-        chiOptions={!players[0]?.isAI ? chiTileOptions ?? undefined : undefined}
-        onChi={!players[0]?.isAI ? handleChiSelect : undefined}
-        tsumoOption={!players[0]?.isAI ? tsumoOption : false}
-        onTsumo={!players[0]?.isAI ? handleTsumo : undefined}
-        onTsumoPass={!players[0]?.isAI ? handleTsumoPass : undefined}
-        ronOption={!players[0]?.isAI ? !!ronCandidate : false}
-        onRon={!players[0]?.isAI ? handleRon : undefined}
-        onRonPass={!players[0]?.isAI ? handleRonPass : undefined}
-        playerIsAI={playerIsAI}
-        onToggleAI={togglePlayerAI}
+        players={game.players}
+        dora={game.dora}
+        kyoku={game.kyoku}
+        wallCount={game.wall.length}
+        kyotaku={game.riichiPool}
+        honba={game.honba}
+        onDiscard={game.handleDiscard}
+        isMyTurn={game.turn === 0 && !game.players[0]?.isAI}
+        shanten={game.shanten}
+        lastDiscard={game.lastDiscard}
+        callOptions={!game.players[0]?.isAI ? game.callOptions ?? undefined : undefined}
+        onCallAction={!game.players[0]?.isAI ? game.handleCallAction : undefined}
+        onRiichi={!game.players[0]?.isAI ? game.handleRiichi : undefined}
+        selfKanOptions={!game.players[0]?.isAI ? game.selfKanOptions ?? undefined : undefined}
+        onSelfKan={!game.players[0]?.isAI ? game.handleSelfKan : undefined}
+        chiOptions={!game.players[0]?.isAI ? game.chiTileOptions ?? undefined : undefined}
+        onChi={!game.players[0]?.isAI ? game.handleChiSelect : undefined}
+        tsumoOption={!game.players[0]?.isAI ? game.tsumoOption : false}
+        onTsumo={!game.players[0]?.isAI ? game.handleTsumo : undefined}
+        onTsumoPass={!game.players[0]?.isAI ? game.handleTsumoPass : undefined}
+        ronOption={!game.players[0]?.isAI ? !!game.ronCandidate : false}
+        onRon={!game.players[0]?.isAI ? game.handleRon : undefined}
+        onRonPass={!game.players[0]?.isAI ? game.handleRonPass : undefined}
+        playerIsAI={game.playerIsAI}
+        onToggleAI={game.togglePlayerAI}
       />
-      <div className="mt-2">{message}</div>
-      <button className="px-2 py-1 bg-gray-200 rounded" onClick={handleDownloadLog}>
+      <div className="mt-2">{game.message}</div>
+      <button className="px-2 py-1 bg-gray-200 rounded" onClick={game.handleDownloadLog}>
         ログダウンロード
       </button>
       <div className="my-2 flex items-center gap-2">
@@ -1017,8 +56,8 @@ const handleCallAction = (action: MeldType | 'pass') => {
           id="preset"
           aria-label="プリセット"
           className="border px-2 py-1"
-          value={preset}
-          onChange={e => setPreset(e.target.value as keyof typeof boardPresets)}
+          value={game.preset}
+          onChange={e => game.setPreset(e.target.value as any)}
         >
           <option value="basic">基本形</option>
           <option value="multiCalls">複数鳴き</option>
@@ -1029,36 +68,38 @@ const handleCallAction = (action: MeldType | 'pass') => {
       <textarea
         aria-label="盤面入力"
         className="w-full h-40 p-1 border font-mono"
-        value={boardInput}
-        onChange={e => setBoardInput(e.target.value)}
+        value={game.boardInput}
+        onChange={e => game.setBoardInput(e.target.value)}
       />
-      <button className="px-2 py-1 bg-gray-200 rounded" onClick={handleLoadBoard}>
+      <button className="px-2 py-1 bg-gray-200 rounded" onClick={game.handleLoadBoard}>
         盤面読み込み
       </button>
-      {winResult && (
+      {game.winResult && (
         <WinResultModal
-          {...winResult}
-          nextLabel={kyokuRef.current >= maxKyoku ? '結果発表へ' : undefined}
+          {...game.winResult}
+          nextLabel={game.kyoku >= maxKyoku ? '結果発表へ' : undefined}
           onNext={() => {
-            const dealerWon = winResult.winner === 0;
-            setWinResult(null);
-            nextKyoku(dealerWon);
+            const dealerWon = game.winResult!.winner === 0;
+            game.setWinResult(null);
+            game.nextKyoku(dealerWon);
           }}
         />
       )}
-      {roundResult && (
+      {game.roundResult && (
         <RoundResultModal
-          results={roundResult.results}
-          nextLabel={kyokuRef.current >= maxKyoku ? '結果発表へ' : undefined}
+          results={game.roundResult.results}
+          nextLabel={game.kyoku >= maxKyoku ? '結果発表へ' : undefined}
           onNext={() => {
-            setRoundResult(null);
-            nextKyoku(true);
+            game.setRoundResult(null);
+            game.nextKyoku(true);
           }}
         />
       )}
-      {phase === 'end' && (
-        <FinalResultModal players={players} onReplay={handleRestart} />
+      {game.phase === 'end' && (
+        <FinalResultModal players={game.players} onReplay={game.handleRestart} />
       )}
     </div>
   );
 };
+
+export { maxKyokuForLength };
